Moved command to subcommand
